Move all enums.ts files to types.ts
