+++ config/yo/generator-component/app/templates/types.ts
