+++ src/components/Avatar/types.ts
