+++ src/components/FixedGrid/types.ts
