+++ src/components/Heading/types.ts
