+++ src/components/Icon/index.ts
@@ -1,3 +1,3 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export * from './enums';
+export * from './types';
 export { default as BaseIcon, IconProps } from './BaseIcon';
