+++ src/components/Image/types.ts
