+++ src/components/Spinner/Spinner.tsx
@@ -9,7 +9,7 @@ import {
 import { BaseComponentProps } from '../../util/BaseComponent/props';
 import Text, { TextProps, TextColor, TextSize } from '../Text';
 import ScreenreaderText from '../ScreenreaderText';
-import { SpinnerSize } from './enums';
+import { SpinnerSize } from './types';
 import './Spinner.css';
 
 export { SpinnerSize };
