+++ src/components/Spinner/types.ts
