+++ src/components/Text/types.ts
