+++ src/util/styles/gutters.ts
@@ -1,5 +1,5 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-import { GutterSize } from '../../components/FixedGrid/enums';
+import { GutterSize } from '../../components/FixedGrid/types';
 
 export const gutterSize: { [size in GutterSize]: string } = {
   none: '0',
